+++ tests/TestCase/DatabaseBackupPluginTest.php
@@ -17,7 +17,7 @@ namespace DatabaseBackup\Test\TestCase;
 
 use App\Application;
 use Cake\Console\CommandCollection;
-use DatabaseBackup\Plugin as DatabaseBackup;
+use DatabaseBackup\DatabaseBackupPlugin;
 use DatabaseBackup\TestSuite\TestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
@@ -25,10 +25,10 @@ use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 
 /**
- * PluginTest.
+ * DatabaseBackupPluginTest.
  */
-#[CoversClass(DatabaseBackup::class)]
-class PluginTest extends TestCase
+#[CoversClass(DatabaseBackupPlugin::class)]
+class DatabaseBackupPluginTest extends TestCase
 {
     /**
      * Tests that the application console includes all expected commands.
@@ -51,7 +51,7 @@ class PluginTest extends TestCase
 
         $CommandCollection = new CommandCollection();
         $App = new Application(CONFIG);
-        $App->addPlugin(DatabaseBackup::class);
+        $App->addPlugin(DatabaseBackupPlugin::class);
         $App->pluginConsole($CommandCollection);
         $availableCommands = $CommandCollection->keys();